+++ src/main/java/com/dominest/dominestbackend/api/category/controller/CategoryController.java
@@ -4,9 +4,9 @@ import com.dominest.dominestbackend.api.category.request.CategoryCreateRequest;
 import com.dominest.dominestbackend.api.category.request.CategoryUpdateRequest;
 import com.dominest.dominestbackend.api.category.response.CategoryListDto;
 import com.dominest.dominestbackend.api.common.ResTemplate;
-import com.dominest.dominestbackend.domain.category.Category;
-import com.dominest.dominestbackend.domain.category.repository.CategoryRepository;
-import com.dominest.dominestbackend.domain.category.service.CategoryService;
+import com.dominest.dominestbackend.domain.post.component.category.Category;
+import com.dominest.dominestbackend.domain.post.component.category.repository.CategoryRepository;
+import com.dominest.dominestbackend.domain.post.component.category.service.CategoryService;
 import com.dominest.dominestbackend.domain.user.User;
 import com.dominest.dominestbackend.domain.user.service.UserService;
 import com.dominest.dominestbackend.global.exception.ErrorCode;
