--- src/main/java/com/dominest/dominestbackend/domain/category/repository/CategoryRepository.java
@@ -1,7 +1,0 @@
-package com.dominest.dominestbackend.domain.category.repository;
-
-import com.dominest.dominestbackend.domain.category.Category;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CategoryRepository extends JpaRepository<Category, Long> {
-}
