--- src/main/java/com/dominest/dominestbackend/domain/categorytype/Type.java
@@ -1,9 +1,0 @@
-package com.dominest.dominestbackend.domain.categorytype;
-
-import lombok.Getter;
-
-@Getter
-public enum Type {
-    PHOTO_UPLOAD,
-    WRITE_POST
-}
