+++ src/main/java/com/dominest/dominestbackend/domain/post/component/category/component/categorytype/Type.java
@@ -1,0 +1,9 @@
+package com.dominest.dominestbackend.domain.post.component.category.component.categorytype;
+
+import lombok.Getter;
+
+@Getter
+public enum Type {
+    PHOTO_UPLOAD,
+    WRITE_POST
+}
