+++ src/main/java/com/dominest/dominestbackend/domain/post/component/category/repository/CategoryRepository.java
@@ -1,0 +1,7 @@
+package com.dominest.dominestbackend.domain.post.component.category.repository;
+
+import com.dominest.dominestbackend.domain.post.component.category.Category;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CategoryRepository extends JpaRepository<Category, Long> {
+}
